requirements.txt
# Recall AI Dependencies
# Compatible versions to avoid conflicts with other packages

# Core Bot Framework
python-telegram-bot==21.5
python-dotenv==1.0.1

# AI Services - Multiple Provider Support
openai==1.51.0
google-generativeai==0.8.3
anthropic==0.34.2
httpx>=0.27.0

# Web Framework for Health Checks
aiohttp==3.10.11

# Databases
pymongo==4.8.0
qdrant-client==1.11.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2

# Image Processing
Pillow>=10.4.0

# Audio Processing
pydub==0.25.1
SpeechRecognition==3.10.4

# Web Processing
requests>=2.32.5
beautifulsoup4==4.12.3

# Data Processing
numpy<2.0
tiktoken==0.8.0

# Utilities
aiofiles==24.1.0
tqdm==4.66.5
